Fix to make tracker breadcrumb trains work
I was always deeply uncomfortable with an "is not" memory check here.

The only explanation I can find for not detecting the breadcrumb trail is horusHunt.name never "is" new.name ... even when they are the same name.

The result is the previous_death and the new death are always the same number in the later if check.